requirements.txt
fastapi[all]==0.115.0
python-multipart==0.0.9
uvicorn[standard]==0.30.0
gunicorn==22.0.0

sqlalchemy==2.0.34
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

redis==5.0.7
celery==5.4.0
flower==2.0.1

httpx==0.27.0
transformers==4.44.2
torch==2.4.1
scikit-learn==1.5.1
nltk==3.9.1

pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

ruff==0.6.2
black==24.8.0
isort==5.13.2

pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
asgiref==3.8.1
bcrypt==4.0.1
greenlet==3.0.3
